Add files via upload
This exercise aims to visualise relationships and outliers from a synthetic dataset of historic vehicle accidents on UK roads, spanning five years. The findings may not seem realistic, as the dataset is generated and does not correspond to real life. The following pages will display five relationships found in the data and two outliers.


A scatterplot is used to visualise this outlier. Moreover, there is a clear relationship between Severity and Damage which are continuous variables. As seen in fig.1 “Se-verity” and “Damage” are connected. The outliers are the points around the line, which do not obey the relationship between the two variables. This is a global outlier in table “Damage”.